+++ app/index.ts
@@ -1,2 +1,3 @@
 // Entry point of the library.
 export {expandMfText} from './converter/mftext';
+export {setWorkingDom} from './utils/dom';