+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - The user interface was restyled to make it less intrusive to the contents.
 - The search field is now functional.
+- `expandMfText` now can be used in a Node.js environment.
 
 ## [0.0.5] - 2018-12-30